- Move end shipper to main namespace - Update unit tests - fail because endpoint is not live
